requirements.txt
fastapi 
uvicorn 
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-auth 
google-api-python-client
gspread
google-cloud-storage
pydub
python-multipart
cryptography
user-agents
pandas
